Canvas->dynamic and overlay for responsiveness
Moves the tissue mask and edge mask display to a transparent canvas
that is updated on command (as opposed to every time a new camera frame
is available). This was done because displaying the masks is a expensive
process and unnecessary to compute for each new frame. Therefore moving
it to a frame updated on command speeds up the responsiveness of the GUI